+++ src/FakeXrmEasy.Plugins/IXrmFakedContextPipelineExtensions.cs
@@ -327,31 +327,6 @@ namespace FakeXrmEasy.Pipeline
             return sdkMessageProcessingStepId;
         }
 
-        internal static void ExecutePipelineStage(this IXrmFakedContext context, 
-            PipelineStageExecutionParameters parameters,
-            object target)
-        {
-            var plugins = context.GetPluginStepsForOrganizationRequest(parameters.RequestName, parameters.Stage, parameters.Mode, parameters.Request);
-            if(plugins == null)
-                return;
-
-            if (target == null)
-            {
-                target = GetTargetForRequest(parameters.Request);
-            }
-
-            if (target is Entity)
-            {
-                parameters.Entity = target as Entity;
-            }
-            else if (target is EntityReference)
-            {
-                parameters.EntityReference = target as EntityReference;
-            }
-
-            context.ExecutePipelineStage(parameters);
-        }
-
         internal static IEnumerable<PluginStepDefinition> GetPluginStepsForOrganizationRequestWithRetrieveMultiple(this IXrmFakedContext context, string requestName, ProcessingStepStage stage, ProcessingStepMode mode, OrganizationRequest request)
         {
             var target = GetTargetForRequest(request);
@@ -392,28 +367,44 @@ namespace FakeXrmEasy.Pipeline
 
                 return context.GetStepsForStage(requestName, stage, mode, (Entity)Activator.CreateInstance(entityType));
             }
-
-            return null;
+            else
+            {
+                //Possibly a custom api execution
+                return context.GetStepsForStage(requestName, stage, mode, null);
+            }
         }
 
-        private static void ExecutePipelineStage(this IXrmFakedContext context, PipelineStageExecutionParameters parameters)
+        internal static void ExecutePipelineStage(this IXrmFakedContext context, PipelineStageExecutionParameters parameters)
         {
-            if(parameters.Entity != null)
+            var plugins = context.GetPluginStepsForOrganizationRequest(parameters.RequestName, parameters.Stage, parameters.Mode, parameters.Request);
+            if (plugins == null)
+                return;
+
+            var target = GetTargetForRequest(parameters.Request);
+
+            if (target != null)
             {
-                var plugins = context.GetStepsForStage(parameters.RequestName, parameters.Stage, parameters.Mode, parameters.Entity);
-                context.ExecutePipelinePlugins(plugins, parameters.Entity, parameters.PreEntitySnapshot, parameters.PostEntitySnapshot, parameters.Response);
-            }
-            else if(parameters.EntityReference != null)
-            {
-                var entityType = context.FindReflectedType(parameters.EntityReference.LogicalName);
-                if (entityType == null)
+                var entityTarget = target as Entity;
+                if (entityTarget != null)
                 {
-                    return;
+                    context.ExecutePipelinePlugins(plugins, parameters.Request, parameters.PreEntitySnapshot, parameters.PostEntitySnapshot, parameters.Response);
                 }
 
-                var plugins = context.GetStepsForStage(parameters.RequestName, parameters.Stage, parameters.Mode, (Entity)Activator.CreateInstance(entityType));
-                context.ExecutePipelinePlugins(plugins, parameters.EntityReference, parameters.PreEntitySnapshot, parameters.PostEntitySnapshot, parameters.Response);
-            }           
+                var entityReferenceTarget = target as EntityReference;
+                if (entityReferenceTarget != null)
+                {
+                    var entityType = context.FindReflectedType(entityReferenceTarget.LogicalName);
+                    if (entityType == null)
+                    {
+                        return;
+                    }
+                    context.ExecutePipelinePlugins(plugins, parameters.Request, parameters.PreEntitySnapshot, parameters.PostEntitySnapshot, parameters.Response);
+                }
+            }
+            else
+            {
+                context.ExecutePipelinePlugins(plugins, parameters.Request, null, null, parameters.Response);
+            }
         }
 
         /// <summary>
@@ -421,13 +412,13 @@ namespace FakeXrmEasy.Pipeline
         /// </summary>
         /// <param name="context"></param>
         /// <param name="pluginSteps"></param>
-        /// <param name="target"></param>
+        /// <param name="organizationRequest"></param>
         /// <param name="previousValues"></param>
         /// <param name="resultingAttributes"></param>
         /// <param name="organizationResponse">The organization response that triggered this plugin execution</param>
         private static void ExecutePipelinePlugins(this IXrmFakedContext context, 
                                                     IEnumerable<PluginStepDefinition> pluginSteps, 
-                                                    object target, 
+                                                    OrganizationRequest organizationRequest,
                                                     Entity previousValues, 
                                                     Entity resultingAttributes,
                                                     OrganizationResponse organizationResponse)
@@ -454,10 +445,7 @@ namespace FakeXrmEasy.Pipeline
                 pluginContext.Mode = (int) pluginStep.Mode;
                 pluginContext.Stage = (int) pluginStep.Stage;
                 pluginContext.MessageName = pluginStep.MessageName;
-                pluginContext.InputParameters = new ParameterCollection
-                {
-                    { "Target", target }
-                };
+                pluginContext.InputParameters = organizationRequest.Parameters;
                 pluginContext.OutputParameters = organizationResponse != null ? organizationResponse.Results : new ParameterCollection();
                 pluginContext.PreEntityImages = GetEntityImageCollection(preImageDefinitions, previousValues);
                 pluginContext.PostEntityImages = GetEntityImageCollection(postImageDefinitions, resultingAttributes);
@@ -466,7 +454,7 @@ namespace FakeXrmEasy.Pipeline
 
                 if(isAuditEnabled)
                 {
-                    context.AddPluginStepAuditDetails(pluginMethod, pluginContext, pluginStep, target);
+                    context.AddPluginStepAuditDetails(pluginMethod, pluginContext, pluginStep);
                 }
             }
         }
@@ -474,8 +462,7 @@ namespace FakeXrmEasy.Pipeline
         private static void AddPluginStepAuditDetails(this IXrmFakedContext context, 
                                 MethodInfo pluginMethod, 
                                 XrmFakedPluginExecutionContext pluginContext, 
-                                PluginStepDefinition pluginStep, 
-                                object target)
+                                PluginStepDefinition pluginStep)
         {
             var pluginType = pluginMethod.GetGenericArguments()[0];
             var pluginStepAuditDetails = new PluginStepAuditDetails()
@@ -488,11 +475,16 @@ namespace FakeXrmEasy.Pipeline
                 OutputParameters = pluginContext.OutputParameters
             };
             
-            if (target is Entity) 
-                pluginStepAuditDetails.TargetEntity = (Entity) target;
+            if(pluginContext.InputParameters.ContainsKey("Target"))
+            {
+                var target = pluginContext.InputParameters["Target"];
+                if (target is Entity)
+                    pluginStepAuditDetails.TargetEntity = (Entity)target;
+
+                if (target is EntityReference)
+                    pluginStepAuditDetails.TargetEntityReference = (EntityReference)target;
+            }
             
-            if (target is EntityReference)
-                pluginStepAuditDetails.TargetEntityReference = (EntityReference)target;
 
 
             var pluginStepAudit = context.GetProperty<IPluginStepAudit>() as PluginStepAudit;
@@ -608,8 +600,14 @@ namespace FakeXrmEasy.Pipeline
                                                                             ProcessingStepMode mode, 
                                                                             Entity entity)
         {
-            var entityTypeCode = (int?)entity.GetType().GetField("EntityTypeCode")?.GetValue(entity);
-            var entityLogicalName = entity.LogicalName;
+            int? entityTypeCode = null; 
+            string entityLogicalName = null;
+            
+            if(entity != null)
+            {
+                entityTypeCode = (int?) entity.GetType().GetField("EntityTypeCode")?.GetValue(entity);
+                entityLogicalName = entity.LogicalName;
+            }
 
             var pluginSteps = (from step in context.CreateQuery(PluginStepRegistrationEntityNames.SdkMessageProcessingStep)
                                join message in context.CreateQuery(PluginStepRegistrationEntityNames.SdkMessage) on (step[SdkMessageProcessingStepFieldNames.SdkMessageId] as EntityReference).Id equals message.Id
@@ -638,7 +636,7 @@ namespace FakeXrmEasy.Pipeline
             return pluginSteps
                         .Where(ps => ps.EntityLogicalName != null && ps.EntityLogicalName == entityLogicalName || //Matches logical name
                                         ps.EntityTypeCode != null && ps.EntityTypeCode.HasValue && ps.EntityTypeCode.Value == entityTypeCode || //Or matches entity type code
-                                        ps.EntityTypeCode == null && ps.EntityLogicalName == null) //Or matches plugins steps with none
+                                        ps.EntityTypeCode == null && ps.EntityLogicalName == null) //Or matches plugins steps with none (Custom Apis)
                         .Where(ps => !ps.FilteringAttributes.Any() || ps.FilteringAttributes.Any(attr => entity.Attributes.ContainsKey(attr))).AsEnumerable();
         }