+++ tests/FakeXrmEasy.Plugins.Tests/CustomApisForTesting/ReverseCustomApiPlugin.cs
@@ -1,0 +1,47 @@
+﻿using Microsoft.Xrm.Sdk;
+using System;
+using System.Linq;
+
+namespace FakeXrmEasy.Plugins.Tests.CustomApisForTesting
+{
+    /// <summary>
+    /// Sample custom api plugin that will reverse a string
+    /// </summary>
+    public class ReverseCustomApiPlugin : IPlugin
+    {
+        public void Execute(IServiceProvider serviceProvider)
+        {
+            // Obtain the tracing service
+            ITracingService tracingService =
+            (ITracingService)serviceProvider.GetService(typeof(ITracingService));
+
+            // Obtain the execution context from the service provider.  
+            IPluginExecutionContext context = (IPluginExecutionContext)
+                serviceProvider.GetService(typeof(IPluginExecutionContext));
+
+            if (context.MessageName.Equals("sample_CustomAPIExample") && context.Stage.Equals(30))
+            {
+
+                try
+                {
+                    string input = (string)context.InputParameters["StringParameter"];
+
+                    if (!string.IsNullOrEmpty(input))
+                    {
+                        //Simply reversing the characters of the string
+                        context.OutputParameters["StringProperty"] = new string(input.Reverse().ToArray());
+                    }
+                }
+                catch (Exception ex)
+                {
+                    tracingService.Trace("Sample_CustomAPIExample: {0}", ex.ToString());
+                    throw new InvalidPluginExecutionException("An error occurred in Sample_CustomAPIExample.", ex);
+                }
+            }
+            else
+            {
+                tracingService.Trace("Sample_CustomAPIExample plug-in is not associated with the expected message or is not registered for the main operation.");
+            }
+        }
+    }
+}