+++ CHANGELOG.md
@@ -2,10 +2,12 @@
 
 ### Changed
 
+- Implement execution of custom apis, as well as registration of extra plugin steps against these apis - https://github.com/DynamicsValue/fake-xrm-easy/issues/50
 - Allow execution of PreImages in prevalidation : https://github.com/DynamicsValue/fake-xrm-easy/issues/71
  
 ### Added
 
+- Added execution of fake custom api's
 - Added validation of preimages and postimages during plugin step registration - https://github.com/DynamicsValue/fake-xrm-easy/issues/33
 - Added custom function to support registration of plugin steps via a custom reflection function (CustomPluginStepDiscoveryFunction) - https://github.com/DynamicsValue/fake-xrm-easy/issues/11
 - First implementation (preview) of automatic registration of plugin steps via reflection - https://github.com/DynamicsValue/fake-xrm-easy/issues/11